+++ pyfsmlib/state_machine.py
