+++ tests/testFSM.py
@@ -1,5 +1,5 @@
 import unittest
-import pyfsm.state_machine as fsm
+import pyfsmlib.state_machine as fsm
 
 
 def initStateMachine(obj):
